Fix `node-fibers debug`
node's debugger is built as a node process which spawns another node
process. There was an issue where the spawned node process (which runs
the client code) was not getting shimmed, because coroutine.cc attempts
to clean up its environment. So I moved the environment cleanup code
to fibers.js which actually simplified some of the logic there.

Closes gh-25